fix the entry
