+++ pyproject.toml
@@ -14,6 +14,9 @@ dependencies = [
     "pymysql>=1.1.1",
 ]
 
+[project.scripts]
+tidb-mcp-server = "pytidb.ext.mcp:main"
+
 [project.optional-dependencies]
 mcp = [
     "mcp[cli]>=1.6.0",
